Move compilation files to build-system/compile/
