+++ build-system/compile/compile.js
@@ -29,11 +29,11 @@ const {
   handleTypeCheckError,
 } = require('./closure-compile');
 const {checkTypesNailgunPort, distNailgunPort} = require('../tasks/nailgun');
-const {CLOSURE_SRC_GLOBS, SRC_TEMP_DIR} = require('../sources');
+const {CLOSURE_SRC_GLOBS, SRC_TEMP_DIR} = require('./sources');
 const {isTravisBuild} = require('../travis');
 const {shortenLicense, shouldShortenLicense} = require('./shorten-license');
 const {singlePassCompile} = require('./single-pass');
-const {VERSION: internalRuntimeVersion} = require('../internal-version');
+const {VERSION: internalRuntimeVersion} = require('./internal-version');
 
 const isProdBuild = !!argv.type;
 const queue = [];
