+++ build-system/pr-check/build-targets.js
@@ -45,9 +45,9 @@ const targetMatchers = [
     targets: ['BABEL_PLUGIN', 'RUNTIME'], // Test the runtime for babel plugin changes.
     func: file => {
       return (
-        file == 'build-system/internal-version.js' ||
         file == 'build-system/babel-plugins/log-module-metadata.js' ||
         file == 'build-system/babel-plugins/static-template-metadata.js' ||
+        file == 'build-system/compile/internal-version.js' ||
         file == 'build-system/compile/log-messages.js' ||
         file == 'build-system/tasks/babel-plugin-tests.js' ||
         file.startsWith('build-system/babel-plugins/')
